+++ docker-compose.yml
@@ -27,7 +27,7 @@ services:
       - pik-network
 
   pik-nlp:
-    image: huseongyoon/piknlp-server:v1.3.0
+    image: huseongyoon/piknlp-server:1.5.0
     container_name: pik-nlp
     # deploy:
     #   resources:
@@ -44,7 +44,7 @@ services:
       - DEVICE=${DEVICE:-cpu} # 'cuda' 또는 'cpu'
       - API_KEY=${NLP_API_KEY}
       - ENV=${NLP_ENV}
-    command: poetry run uvicorn app.main:app --host 0.0.0.0 --port 18000 --workers 4
+    command: uv run uvicorn app.main:app --host 0.0.0.0 --port 18000 --workers 4
     networks:
       - pik-network
