+++ scripts/initLocationAggregation.js
@@ -1,7 +1,7 @@
 // scripts/initLocationAggregation.js
 require("dotenv").config();
 const mongoose = require("mongoose");
-const Location = require("./models/Location");
+const Location = require("../server/src/models/Location");
 const connectDB = require("./database");
 
 async function init() {
