+++ server/src/routes/categoryRoutes.js
