+++ server/src/routes/locationRoutes.js
