+++ server/src/utils/locationAnalysis.js
