+++ server/src/utils/updateLocationAggregation.js
